refactor: restructure types
