--- src/types/AsyncAction.d.ts
@@ -1,4 +1,0 @@
-/** @format */
-import type { Result } from './Result';
-
-export type AsyncAction<Args extends Array<unknown>, Res> = (...args: Args) => Promise<Res>;
