--- src/types/ErrorHandler.d.ts
@@ -1,3 +1,0 @@
-/** @format */
-
-export type ErrorHandler<Err extends Error | string> = (error: unknown | null) => Err;
