--- src/types/NarrowedResult.d.ts
@@ -1,8 +1,0 @@
-/** @format */
-
-import type { ResultHandler } from './ResultHandler';
-
-export type NarrowedResult<Res> = {
-    result: Res;
-    complete: (handler: ResultHandler<Res>) => void;
-};
