--- src/types/ResultHandler.d.ts
@@ -1,3 +1,0 @@
-/** @format */
-
-export type ResultHandler<Result> = (result: Result) => void;
