+++ src/types/async-handler.d.ts
@@ -1,0 +1,7 @@
+/** @format */
+
+
+export type AsyncHandler<Result, Target> = {
+    Ok: (value: Result) => Target;
+    Err: (error: Error) => never | NoInfer<Target>;
+};
