+++ src/types/async-runner.d.ts
@@ -1,0 +1,2 @@
+/** @format */
+export type AsyncRunner<Result> = (...args: any[]) => Promise<Result>;
