+++ src/types/error-like.d.ts
