+++ src/types/index.d.ts
@@ -1,7 +1,5 @@
 /** @format */
-export type { AsyncAction } from './AsyncAction';
-export type { ErrorHandler } from './ErrorHandler';
-export type { NarrowedResult } from './NarrowedResult';
-export type { Result } from './Result';
-export type { ResultHandler } from './ResultHandler';
-export type { ErrorLike } from './ErrorLike';
+
+export type { ErrorLike } from './error-like';
+export type { AsyncRunner } from './async-runner';
+export type { AsyncHandler } from './async-handler';
